+++ CHANGELOG.rst
@@ -1,6 +1,13 @@
 Changelog
 #########
 
+2.0.0
+*****
+
+- changed most class members from protected to private
+- cs fixes, added codestyle checks
+
+
 1.1.0
 *****